order algo long and short --/short/requirements.txt
requests
python-binance
watchdog
python-telegram-bot